Fixed changes in behaviour after using wrapper
Some of the SnackbarWrapper settings shouldn't be called if that field was never initialised in the builder, such as background colour (no colour) and icon (no icon).

Sanitised the action click listener, builder pattern means user might do the action click listener through a callback, whereas Snackbar hides action if this click listener is null.

Added a default action text color, similarly to the message text colur.

Fixed naming of the methods in wrapper to change icon margin to setIconMargin and setIconMarginRes. Previously the default method took a resource and other method had Pixels prefix, which easily resulted in bugs.

Added @NonNull for the appendMessage method on wrapper, as it causes it to crash if you pass null.